pyproject.toml
[tool.poetry]
name = "lms-project"
version = "0.1.0"
description = ""
authors = ["Pinchuk Sergey <Pinboyer@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2"
djangorestframework = "^3.15.2"
django-stubs = "^5.1.0"
psycopg2-binary = "^2.9.10"
pillow = "^11.0.0"
python-dotenv = "^1.0.1"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
ipython = "^8.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
